refactor: switch 구문 메핑 구조로 변경
